`yaml` reader was broken for hyperdrive patch models
